requirements.txt
streamlit==1.37.0
sentence-transformers==3.0.1
faiss-cpu==1.11.0
numpy==1.26.4
google-generativeai==0.7.2
